Six Sigma Tools for Quality and Process Improvement

Six Sigma Tools for Quality and Process Improvement

Details

This package contains functions and utilities to perform Statistical Analyses in the Six Sigma way. Through the DMAIC cycle (Define, Measure, Analyze, Improve, Control), you can manage several Quality Management studies: Gage R&R, Capability Analysis, Control Charts, Loss Function Analysis, etc. Data frames used in "Six Sigma with R" (Springer, 2012) are also included in the package. Use the package to perform Six Sigma Methodology tasks, throughout its breakthrough strategy: Define, Measure, Analyze, Improve, Control (DMAIC)
Define: Process Map (ss.pMap), Cause and effect Diagram (ss.ceDiag);
Measure: Gage R&R study (ss.rr); Capability Analysis (ss.study.ca); Loss Function Analysis (ss.lfa)
Analyze: Confidence Intervals (ss.ci)
Control: Moving Average Control Chart
Soon: further functions

Note

The current version includes Loss Function Analysis, Gage R&R Study, confidence intervals, Process Map and Cause-and-Effect diagram. We plan to regularly upload updated versions, with new functions and improving those previously deployed. The subsequent versions will cover tools for the whole cycle:

Define

Measure

Analyze

Improve

Control

References

Allen, T. T. (2010) Introduction to Engineering Statistics and Lean Six Sigma - Statistical Quality Control and Design of Experiments and Systems (Second Edition ed.). London: Springer.

Box, G. (1991). Teaching engineers experimental design with a paper helicopter. Report 76, Center for Quality and Productivity Improvement. University of Wisconsin.

Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012. Six Sigma with R. Statistical Engineering for Process Improvement, Use R!, vol. 36. Springer, New York. https://link.springer.com/book/10.1007/978-1-4614-3652-2/.

Cano, Emilio L., Moguerza, Javier M. and Prieto Corcoba, Andrés. 2015. Quality Control with R. An ISO Standards approach, Use R!, Springer, New York.

Chambers, J. M. (2008) Software for data analysis. Programming with R New York: Springer.

Montgomery, DC (2008) Introduction to Statistical Quality Control (Sixth Edition). New York: Wiley&Sons

Wikipedia, https://en.wikipedia.org/wiki/Six_Sigma/

See also

Author

Emilio L. Cano, Javier M. Moguerza, Mariano Prieto Corcoba and Andrés Redchuk;

Maintainer: Emilio L. Cano emilio.lopez@urjc.es