Capability Indices

Compute the Capability Indices of a process, Z (Sigma Score), \(C_p\) and \(C_{pk}\).

ss.ca.cp(x, LSL = NA, USL = NA, LT = FALSE, f.na.rm = TRUE, 
  ci = FALSE, alpha = 0.05)

ss.ca.cpk(x, LSL = NA, USL = NA, LT = FALSE, f.na.rm = TRUE, 
  ci = FALSE, alpha = 0.05)

ss.ca.z(x, LSL = NA, USL = NA, LT = FALSE, f.na.rm = TRUE)

Arguments

x

A vector with the data of the process performance

LSL

Lower Specification Limit

USL

Upper Specification Limit

LT

Long Term data (TRUE/FALSE). Not used for the moment

f.na.rm

Remove NA data (TRUE/FALSE)

ci

If TRUE computes a Confidence Interval

alpha

Type I error (\(\alpha\)) for the Confidence Interval

Value

A numeric value for the index, or a vector with the limits of the Confidence Interval

References

Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andres. 2012. Six Sigma with R. Statistical Engineering for Process Improvement, Use R!, vol. 36. Springer, New York. https://link.springer.com/book/10.1007/978-1-4614-3652-2/.

Montgomery, DC (2008) Introduction to Statistical Quality Control (Sixth Edition). New York: Wiley&Sons

See also

Author

EL Cano

Examples

ss.ca.cp(ss.data.ca$Volume,740, 760)
#> [1] 1.584136
ss.ca.cpk(ss.data.ca$Volume,740, 760)
#> [1] 1.546513
ss.ca.z(ss.data.ca$Volume,740,760)
#> [1] 4.639539