Data set for the printer cartridge example, by region

This data set contains data from a simulated process of printer cartridge filling. The dataframe contains defects by region of each type of cartridge.

data(ss.data.pc.r)

Format

A data frame with 5 observations on the following 4 variables.

pc.regions

a factor with levels region.1 region.2 region.3 region.4 region.5

pc.def.a

a numeric vector for defects type A

pc.def.b

a numeric vector for defects type B

pc.def

a numeric vector for total defects

Source

See references.

Note

This data set is used in chapter 8 of the book ``Six Sigma with R'' (see References).

References

Cano, Emilio L., Moguerza, Javier M. and Redchuk, Andrés. 2012. Six Sigma with R. Statistical Engineering for Process Improvement, Use R!, vol. 36. Springer, New York. http://www.springer.com/statistics/book/978-1-4614-3651-5.

Examples

data(ss.data.pc.r)
summary(ss.data.pc.r)

#>     pc.regions    pc.def.a       pc.def.b        pc.def     
#>  region.1:1    Min.   :13.0   Min.   :39.0   Min.   : 65.0  
#>  region.2:1    1st Qu.:17.0   1st Qu.:43.0   1st Qu.: 67.0  
#>  region.3:1    Median :37.0   Median :48.0   Median : 77.0  
#>  region.4:1    Mean   :29.2   Mean   :49.4   Mean   : 78.6  
#>  region.5:1    3rd Qu.:38.0   3rd Qu.:54.0   3rd Qu.: 84.0  
#>                Max.   :41.0   Max.   :63.0   Max.   :100.0